+++ src/Rest/Query/Filter/FilterException.php
@@ -9,7 +9,12 @@ class FilterException extends \Exception
     public const CONDITION_PATH_EMPTY = 1;
     public const CONDITION_OPERATOR_UNDEFINED = 2;
     public const RESERVED_FILTER_ITEM_ID = 3;
-    public const CONJUNCTION_UNDEFINED = 4;
+
+    /**
+     * @deprecated Use FilterException::LOGICAL_OPERATOR_UNDEFINED instead
+     */
+    public const CONJUNCTION_UNDEFINED = self::LOGICAL_OPERATOR_UNDEFINED;
+    public const LOGICAL_OPERATOR_UNDEFINED = 4;
     public const ATTRIBUTE_PATH_UNDEFINED = 5;
     public const FILTER_ITEM_INVALID = 6;
     public const CONDITION_PATH_MISSING = 7;